+++ client/src/components/SimpleMap.jsx
